refactor(FILE): move docker build/push plumbing to bin/docker_ops.py
common.py keeps the tag vocabulary, versions and Dockerfile templating it
is imported for; the nine functions that shell out to docker move out with
their bodies unchanged apart from the common. prefix on cross-module calls.
